+++ sqlitemigration/sqlitemigration.go
@@ -81,9 +81,6 @@ type Pool struct {
 	opts   Options
 	cancel context.CancelFunc
 
-	initedMu sync.RWMutex // protects inited
-	inited   map[*sqlite.Conn]struct{}
-
 	ready <-chan struct{} // protects the following fields
 	pool  *sqlitex.Pool
 	err   error
@@ -102,9 +99,6 @@ func NewPool(uri string, schema Schema, opts Options) *Pool {
 		opts:   opts,
 		cancel: cancel,
 		ready:  ready,
-	}
-	if opts.PrepareConn != nil {
-		p.inited = make(map[*sqlite.Conn]struct{})
 	}
 	go func() {
 		defer close(ready)
@@ -153,47 +147,19 @@ func (p *Pool) Get(ctx context.Context) (*sqlite.Conn, error) {
 			ready = true
 		case <-ctx.Done():
 			tick.Stop()
-			return nil, fmt.Errorf("get sqlite conn: %w", ctx.Err())
+			return nil, fmt.Errorf("get sqlite connection: %w", ctx.Err())
 		}
 	}
 	tick.Stop()
 
 	if p.err != nil {
-		return nil, fmt.Errorf("get sqlite conn: %w", p.err)
+		return nil, fmt.Errorf("get sqlite connection: %w", p.err)
 	}
-	conn := p.pool.Get(ctx)
-	if conn == nil {
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("get sqlite conn: %w", err)
-		}
-		return nil, errors.New("get sqlite conn: pool closed")
-	}
-	if err := p.prepare(conn); err != nil {
-		p.pool.Put(conn)
-		return nil, fmt.Errorf("get sqlite conn: %w", err)
+	conn, err := p.pool.Take(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("get sqlite connection: %w", err)
 	}
 	return conn, nil
-}
-
-func (p *Pool) prepare(conn *sqlite.Conn) error {
-	if p.opts.PrepareConn == nil {
-		return nil
-	}
-	p.initedMu.RLock()
-	_, inited := p.inited[conn]
-	p.initedMu.RUnlock()
-	if inited {
-		return nil
-	}
-	if err := p.opts.PrepareConn(conn); err != nil {
-		return fmt.Errorf("prepare connection: %w", err)
-	}
-	// This will not race, since other goroutines will not be able to acquire the
-	// connection from the pool.
-	p.initedMu.Lock()
-	p.inited[conn] = struct{}{}
-	p.initedMu.Unlock()
-	return nil
 }
 
 // Put puts an SQLite connection back into the pool.
@@ -242,26 +208,27 @@ func (p *Pool) open(ctx context.Context, uri string, schema Schema) (*sqlitex.Po
 		}
 
 		pool, err := sqlitex.NewPool(uri, sqlitex.PoolOptions{
-			Flags:    p.opts.Flags,
-			PoolSize: p.opts.PoolSize,
+			Flags:       p.opts.Flags,
+			PoolSize:    p.opts.PoolSize,
+			PrepareConn: p.opts.PrepareConn,
 		})
 		if err != nil {
 			p.opts.OnError.call(err)
 			continue
 		}
-		conn := pool.Get(ctx)
-		if conn == nil {
-			// Canceled.
+
+		conn, err := pool.Take(ctx)
+		if isDone(err) {
 			pool.Close()
 			return nil, errors.New("closed before successful migration")
 		}
-		if err := p.prepare(conn); err != nil {
-			pool.Put(conn)
+		if err != nil {
 			if closeErr := pool.Close(); closeErr != nil {
 				p.opts.OnError.call(fmt.Errorf("close after failed connection preparation: %w", closeErr))
 			}
 			return nil, err
 		}
+
 		err = migrateDB(ctx, conn, schema, p.opts.OnStartMigrate)
 		pool.Put(conn)
 		if err != nil {
@@ -461,4 +428,10 @@ func stepAndReset(stmt *sqlite.Stmt) error {
 		return err
 	}
 	return stmt.Reset()
+}
+
+// isDone reports whether an error indicates cancellation or deadline exceeded
+// from a [context.Context]
+func isDone(err error) bool {
+	return errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded)
 }